requirements.txt
Flask==2.0.1
requests==2.26.0
dnspython
Werkzeug==2.0.1